Cleanup: Make Widget Markup Readable By Consolidating Inputs
Affected Inputs:
- DropdownSelect (previously known as Dropdown)
- RadioSelect (previously known as RadioButton)
- InputBox
