+++ septa-fare-calculator/src/widget/components/InputBox.js
@@ -1,16 +1,19 @@
 import React from 'react';
 
-function InputBox( { value, onChange, type = 'text', placeholder = '', ...rest } ) {
+function InputBox( { label, value, onChange, type = 'text', placeholder = '', ...rest } ) {
   return (
-    <div className="input-box">
-      <input
-        type={ type }
-        value={ value || '' }
-        onChange={ onChange }
-        placeholder={ placeholder }
-        { ...rest }
-      />
-    </div>
+    <>
+      <div className="text-question">{ label }</div>
+      <div className="input-box">
+        <input
+          type={ type }
+          value={ value || '' }
+          onChange={ onChange }
+          placeholder={ placeholder }
+          { ...rest }
+        />
+      </div>
+    </>
   );
 }
