+++ septa-fare-calculator/src/widget/components/index.js
@@ -1,3 +1,3 @@
-export { default as Dropdown } from './Dropdown';
-export { default as RadioButton } from './RadioButton';
+export { default as DropdownSelect } from './DropdownSelect';
+export { default as RadioSelect } from './RadioSelect';
 export { default as InputBox } from './InputBox';
